+++ sdk/metricsadvisor/ai-metrics-advisor/recordings/node/metricsadvisoradministrationclient_datafeed_datafeed/recording_creates_an_azure_data_explorer_feed.js
@@ -1,0 +1,47 @@
+let nock = require('nock');
+
+module.exports.hash = "5dbbc6fb1585472d731178b89744c22e";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint:443', {"encodedQueryParams":true})
+  .post('/metricsadvisor/v1.0/dataFeeds', {"dataSourceType":"AzureDataExplorer","dataFeedName":"js-test-dataExplorerFeed-160072575346104337","granularityName":"Daily","metrics":[{"metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00.000Z","startOffsetInSeconds":0,"maxConcurrency":-1,"minRetryIntervalInSeconds":-1,"stopRetryAfterInSeconds":-1,"needRollup":"NeedRollup","rollUpMethod":"Sum","allUpIdentification":"__CUSTOM_SUM__","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"viewMode":"Private","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","query":"let starttime=datetime(@StartTime); let endtime=starttime"}})
+  .reply(201, "", [
+  'Content-Length',
+  '0',
+  'Location',
+  'https://endpoint/metricsadvisor/v1.0/dataFeeds/c6d3ec1c-4573-4ee3-bada-0231da203c3a',
+  'x-request-id',
+  '633b928d-2289-45ea-981d-48168e01c497',
+  'x-envoy-upstream-service-time',
+  '442',
+  'apim-request-id',
+  '633b928d-2289-45ea-981d-48168e01c497',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains; preload',
+  'x-content-type-options',
+  'nosniff',
+  'Date',
+  'Mon, 21 Sep 2020 22:02:46 GMT'
+]);
+
+nock('https://endpoint:443', {"encodedQueryParams":true})
+  .get('/metricsadvisor/v1.0/dataFeeds/c6d3ec1c-4573-4ee3-bada-0231da203c3a')
+  .reply(200, {"dataFeedId":"c6d3ec1c-4573-4ee3-bada-0231da203c3a","dataFeedName":"js-test-dataExplorerFeed-160072575346104337","metrics":[{"metricId":"72682443-97a4-40b5-9a29-2ca3432b70d6","metricName":"Metric1","metricDisplayName":"Metric1","metricDescription":""},{"metricId":"00117802-197f-4f89-a121-c25f0f540ca0","metricName":"Metric2","metricDisplayName":"Metric2","metricDescription":""}],"dimension":[{"dimensionName":"Dim1","dimensionDisplayName":"Dim1 display"},{"dimensionName":"Dim2","dimensionDisplayName":"Dim2 display"}],"dataStartFrom":"2020-08-21T00:00:00Z","dataSourceType":"AzureDataExplorer","timestampColumn":"","startOffsetInSeconds":0,"maxQueryPerMinute":30,"granularityName":"Daily","granularityAmount":null,"allUpIdentification":"__CUSTOM_SUM__","needRollup":"NeedRollup","fillMissingPointType":"CustomValue","fillMissingPointValue":555,"rollUpMethod":"Sum","rollUpColumns":[],"dataFeedDescription":"","stopRetryAfterInSeconds":-1,"minRetryIntervalInSeconds":-1,"maxConcurrency":-1,"viewMode":"Private","admins":["yumeng@microsoft.com"],"viewers":[],"creator":"yumeng@microsoft.com","status":"Active","createdTime":"2020-09-21T22:02:46Z","isAdmin":true,"actionLinkTemplate":"","dataSourceParameter":{"connectionString":"Server=server.example.net;Encrypt=True;","query":"let starttime=datetime(@StartTime); let endtime=starttime"}}, [
+  'Content-Length',
+  '1345',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'x-request-id',
+  '3a4ead38-fcb1-4eb3-8d3f-5fb84d0a05a7',
+  'x-envoy-upstream-service-time',
+  '131',
+  'apim-request-id',
+  '3a4ead38-fcb1-4eb3-8d3f-5fb84d0a05a7',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains; preload',
+  'x-content-type-options',
+  'nosniff',
+  'Date',
+  'Mon, 21 Sep 2020 22:02:46 GMT'
+]);