+++ sdk/search/search-documents/recordings/node/searchindexerclient_datasourceconnections/recording_throws_error_for_invalid_datasourceconnection_object.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "5fe89803906418440a2c8d22b6894332";
+module.exports.hash = "042821d4c3616b703a08e02f760abe31";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -18,7 +18,7 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Expires',
   '-1',
   'request-id',
-  'c41f1478-4956-4b70-a895-14874e35741c',
+  'c9663ed4-a1f2-4dcd-b43d-1949e748c5cc',
   'elapsed-time',
   '5',
   'OData-Version',
@@ -28,6 +28,6 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=15724800; includeSubDomains',
   'Date',
-  'Wed, 26 Aug 2020 01:12:29 GMT',
+  'Sat, 26 Sep 2020 18:13:51 GMT',
   'Content-Length',
   '112' ]);