+++ sdk/communication/communication-common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 1.0.0-beta.2 (2020-10-06)
+
+Updated `@azure/communication-common` version
+
 ## 1.0.0-beta.1 (2020-09-22)
 
 The Azure Communication Common Client library contains common code such as the CommunicationUserCredential that is being used by other Azure Communication Services.