+++ sdk/communication/communication-sms/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 1.0.0-beta.2 (2020-10-06)
+
+Updated `@azure/communication-sms` version
+
 ## 1.0.0-beta.1 (2020-09-22)
 
 The first preview of the Azure Communication Sms Client has the following features: